started work on sffinfo command.  fixed bug across all paralellized commands if the file position was too big to fit in an int.

+++ aligncommand.cpp
@@ -386,7 +386,7 @@ int AlignCommand::execute(){
 				}
 			}
 			else{
-				vector<int> positions;
+				vector<unsigned long int> positions;
 				processIDS.resize(0);
 				
 				ifstream inFASTA;
@@ -396,7 +396,7 @@ int AlignCommand::execute(){
 				while(!inFASTA.eof()){
 					input = getline(inFASTA);
 					if (input.length() != 0) {
-						if(input[0] == '>'){	long int pos = inFASTA.tellg(); positions.push_back(pos - input.length() - 1);	}
+						if(input[0] == '>'){	unsigned long int pos = inFASTA.tellg(); positions.push_back(pos - input.length() - 1);	}
 					}
 				}
 				inFASTA.close();
@@ -406,7 +406,7 @@ int AlignCommand::execute(){
 				int numSeqsPerProcessor = numFastaSeqs / processors;
 				
 				for (int i = 0; i < processors; i++) {
-					long int startPos = positions[ i * numSeqsPerProcessor ];
+					unsigned long int startPos = positions[ i * numSeqsPerProcessor ];
 					if(i == processors - 1){
 						numSeqsPerProcessor = numFastaSeqs - i * numSeqsPerProcessor;
 					}
