+++ screenseqscommand.h
@@ -23,9 +23,9 @@ public:
 private:
 
 	struct linePair {
-		int start;
+		unsigned long int start;
 		int numSeqs;
-		linePair(long int i, int j) : start(i), numSeqs(j) {}
+		linePair(unsigned long int i, int j) : start(i), numSeqs(j) {}
 	};
 	vector<int> processIDS;   //processid
 	vector<linePair*> lines;
